MathTrig - Fix Phpstan Accomodations
This should be the last of my mass changes to MathTrig. All he Phpstan violations found in baseline which  are part of MathTrig are now fixed and removed from baseline. There were about 20 of these.

+++ src/PhpSpreadsheet/Calculation/MathTrig/RoundUp.php
@@ -20,7 +20,7 @@ class RoundUp
     {
         try {
             $number = Helpers::validateNumericNullBool($number);
-            $digits = Helpers::validateNumericNullSubstitution($digits, null);
+            $digits = (int) Helpers::validateNumericNullSubstitution($digits, null);
         } catch (Exception $e) {
             return $e->getMessage();
         }
